fix: typo on czech translation
fix: subscription progress above 100% for disabled subscriptions
fix: use gd if imagick is not available
fix: bug setting main currency for the first registered user
fix: deprecation message
fix: use first currency on the list of currencies if user has not selected a main currency

+++ endpoints/payments/add.php
@@ -60,16 +60,31 @@ function saveLogo($imageData, $uploadFile, $name, $settings)
         imagepng($image, $tempFile);
         imagedestroy($image);
 
-        $imagick = new Imagick($tempFile);
-        if ($removeBackground) {
-            $fuzz = Imagick::getQuantum() * 0.1; // 10%
-            $imagick->transparentPaintImage("rgb(247, 247, 247)", 0, $fuzz, false);
-        }
-        $imagick->setImageFormat('png');
-        $imagick->writeImage($uploadFile);
+        if (extension_loaded('imagick')) {
+            $imagick = new Imagick($tempFile);
+            if ($removeBackground) {
+                $fuzz = Imagick::getQuantum() * 0.1; // 10%
+                $imagick->transparentPaintImage("rgb(247, 247, 247)", 0, $fuzz, false);
+            }
+            $imagick->setImageFormat('png');
+            $imagick->writeImage($uploadFile);
 
-        $imagick->clear();
-        $imagick->destroy();
+            $imagick->clear();
+            $imagick->destroy();
+        } else {
+            // Alternative method if Imagick is not available
+            $newImage = imagecreatefrompng($tempFile);
+            if ($removeBackground) {
+                imagealphablending($newImage, false);
+                imagesavealpha($newImage, true);
+                $transparent = imagecolorallocatealpha($newImage, 0, 0, 0, 127);
+                imagefill($newImage, 0, 0, $transparent);  // Fill the entire image with transparency
+                imagepng($newImage, $uploadFile);
+                imagedestroy($newImage);
+            }
+            imagepng($newImage, $uploadFile);
+            imagedestroy($newImage);
+        }
         unlink($tempFile);
 
         return true;
@@ -120,13 +135,13 @@ function resizeAndUploadLogo($uploadedFile, $uploadDir, $name)
             $newHeight = $height;
 
             if ($width > $targetWidth) {
-                $newWidth = (int)$targetWidth;
-                $newHeight = (int)(($targetWidth / $width) * $height);
+                $newWidth = (int) $targetWidth;
+                $newHeight = (int) (($targetWidth / $width) * $height);
             }
 
             if ($newHeight > $targetHeight) {
-                $newWidth = (int)(($targetHeight / $newHeight) * $newWidth);
-                $newHeight = (int)$targetHeight;
+                $newWidth = (int) (($targetHeight / $newHeight) * $newWidth);
+                $newHeight = (int) $targetHeight;
             }
 
             $resizedImage = imagecreatetruecolor($newWidth, $newHeight);
